+++ Engine/Common/UserProfile.cs
@@ -10,15 +10,15 @@ namespace JocysCom.VS.AiCompanion.Engine
 		public string Username { get => _Username; set => SetProperty(ref _Username, value); }
 		private string _Username;
 
+		/// <summary>Unique identifier for the account</summary>
+		public string AccountId { get => _AccountId; set => SetProperty(ref _AccountId, value); }
+		private string _AccountId;
+
 		public string Email { get => _Email; set => SetProperty(ref _Email, value); }
 		private string _Email;
 
 		public ApiServiceType ServiceType { get => _ServiceType; set => SetProperty(ref _ServiceType, value); }
 		private ApiServiceType _ServiceType;
-
-		/// <summary>Avatar URL</summary>
-		public string AvatarUrl { get => _AvatarUrl; set => SetProperty(ref _AvatarUrl, value); }
-		private string _AvatarUrl;
 
 		/// <summary>Access token.</summary>
 		[XmlIgnore, JsonIgnore]