+++ Engine/Controls/AuthControl.xaml.cs
@@ -1,15 +1,13 @@
 ﻿using Azure.Identity;
 using Azure.Security.KeyVault.Secrets;
-using JocysCom.ClassLibrary;
 using JocysCom.ClassLibrary.Controls;
-using Microsoft.Identity.Client;
+using JocysCom.VS.AiCompanion.Engine.Security;
 using System;
-using System.Linq;
+using System.IdentityModel.Tokens.Jwt;
 using System.Net.Http;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Media.Imaging;
 
 namespace JocysCom.VS.AiCompanion.Engine.Controls
 {
@@ -18,9 +16,6 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 	/// </summary>
 	public partial class AuthControl : UserControl
 	{
-		private IPublicClientApplication _pca;
-		private IAccount _account;
-
 		public AuthControl()
 		{
 			InitializeComponent();
@@ -29,134 +24,28 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 		private async void SignInButton_Click(object sender, RoutedEventArgs e)
 		{
 			LogPanel.Clear();
-			var scopes = new string[] { "User.Read" };
-			AuthenticationResult result;
-			try
-			{
-				// The application requests User.Read permission,
-				// which allows it to sign in the user and read their basic profile information.
-
-				/*
-
-				When trying to sign-in for the first time into `user.name@company.com` Azure account user will see
-				"Approval Required" dialog if appliction has not been certified or previously approved within the "Company.com" organization's Azure AD.
-				"Company.com" administrator must approve these permissions for the application:
-					- Sign in and read the user profile:
-					  This allows the app to sign in users and read their basic profile information.
-					- Maintain access to data the user has given it access to:
-				      This allows the app to retain access to the granted resources
-					  without needing further explicit user consent repeatedly.
-
-				Who Needs to Approve?
-				IT or Domain Administrator of `company.com` domain.
-
-				AI Companion asks for `User.Read` permission that typically grants the application the ability to:
-
-					- Access the user’s basic profile info.
-					- Access the user’s email address.
-					- Access the user’s display name.
-
-					Applciaiton This data will be stored on the local PC.
-
-				*/
-
-				_account = (await _pca.GetAccountsAsync()).FirstOrDefault();
-				// Get result that includes access tokens that grant the application the rights to fetch user information.
-				result = await _pca.AcquireTokenSilent(scopes, _account).ExecuteAsync();
-			}
-			catch (MsalUiRequiredException)
-			{
-				result = await _pca.AcquireTokenInteractive(scopes).ExecuteAsync();
-			}
-			catch
-			{
-				return;
-			}
-			_account = result.Account;
-			UserName.Text = result.Account.Username;
-			var avatarUrl = await LoadUserAvatar(result.AccessToken);
-			SaveUserProfile(result, avatarUrl);
+			await Global.Security.SignIn();
+			await LoadUserProfile();
 		}
 
-
-		/// <summary>
-		/// Get user profile with the access token.
-		/// </summary>
-		public OperationResult<UserProfile> GetProfile()
+		private async void SignOutButton_Click(object sender, RoutedEventArgs e)
 		{
-			var profile = Global.AppSettings.UserProfiles.FirstOrDefault(p => p.ServiceType == ApiServiceType.Azure);
-			if (profile == null)
-				return new OperationResult<UserProfile>(new Exception("Profile not found. Please log-in."));
-			if (string.IsNullOrEmpty(profile.AccessToken))
-				return new OperationResult<UserProfile>(new Exception("No valid user profile or access token found."));
-			return new OperationResult<UserProfile>(profile);
-		}
-
-
-		private void SaveUserProfile(AuthenticationResult result, string avatarUrl)
-		{
-			var account = result.Account;
-			// Save User profile.
-			var profile = Global.AppSettings.UserProfiles
-				.FirstOrDefault(x => x.ServiceType == ApiServiceType.Azure && x.Username == account.Username);
-			if (profile == null)
-			{
-				profile = new UserProfile()
-				{
-					ServiceType = ApiServiceType.Azure,
-					Username = account.Username,
-				};
-				Global.AppSettings.UserProfiles.Add(profile);
-			}
-			profile.AccessToken = result.AccessToken;
-			profile.AvatarUrl = avatarUrl;
-		}
-
-		private async Task LoadUserProfile()
-		{
-			// Load saved user profile
-			var profileResult = GetProfile();
-			if (!profileResult.Success)
-			{
-				LogPanel.Add(string.Join("\r\n", profileResult.Errors));
-				return;
-			}
-			var profile = profileResult.Result;
-			UserName.Text = profile.Username;
-			await LoadUserAvatar(profile.AvatarUrl);
-		}
-
-		private async Task<string> LoadUserAvatar(string accessToken)
-		{
-			using (var httpClient = new HttpClient())
-			{
-				httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", accessToken);
-				var response = await httpClient.GetAsync("https://graph.microsoft.com/v1.0/me/photo/$value");
-				if (response.IsSuccessStatusCode)
-				{
-					var avatarUrl = response.RequestMessage.RequestUri.ToString();
-					var stream = await response.Content.ReadAsStreamAsync();
-					var bitmap = new BitmapImage();
-					bitmap.BeginInit();
-					bitmap.StreamSource = stream;
-					bitmap.CacheOption = BitmapCacheOption.OnLoad;
-					bitmap.EndInit();
-					UserAvatar.Source = bitmap;
-					return avatarUrl;
-				}
-			}
-			return null;
+			var success = await Global.Security.SignOut();
+			LogPanel.Add(
+				success
+				? "User signed out successfully.\r\n"
+				: "No user is currently signed in.\r\n"
+				);
+			UserName.Text = string.Empty;
+			UserAvatar.Source = null;
 		}
 
 		private async void This_Loaded(object sender, RoutedEventArgs e)
 		{
+			// Allows to run mehod once when control is created.
 			if (ControlsHelper.AllowLoad(this))
 			{
-				_pca = PublicClientApplicationBuilder.Create(Global.AppSettings?.ClientAppId)
-					.WithRedirectUri("http://localhost")
-					.WithDefaultRedirectUri()
-					.Build();
-
+				await Global.Security.LoadCurrentAccount();
 				await LoadUserProfile();
 			}
 		}
@@ -166,12 +55,11 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 		/// </summary>
 		public async Task FetchAzureInformation()
 		{
-			LogPanel.Clear();
 			// Load saved user profile
-			var profileResult = GetProfile();
+			var profileResult = Global.Security.GetProfile();
 			if (!profileResult.Success)
 			{
-				LogPanel.Add(string.Join("\r\n", profileResult.Errors));
+				LogPanel.Add(string.Join("\r\n", profileResult.Errors) + "\r\n");
 				return;
 			}
 			var profile = profileResult.Result;
@@ -183,15 +71,15 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 					var response = await httpClient.GetAsync(TestTextBox.Text);
 					if (!response.IsSuccessStatusCode)
 					{
-						LogPanel.Add("API request failed with status code: " + response.StatusCode);
+						LogPanel.Add($"{response}\r\n");
 						return;
 					}
 					var contents = await response.Content.ReadAsStringAsync();
-					LogPanel.Add(contents);
+					LogPanel.Add($"{contents}\r\n");
 				}
 				catch (System.Exception ex)
 				{
-					LogPanel.Add(ex.ToString());
+					LogPanel.Add($"{ex}\r\n");
 				}
 			}
 		}
@@ -215,12 +103,12 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 				var client = new SecretClient(new Uri(kvUri), new ClientSecretCredential(tenantId, clientId, clientSecret));
 				// Retrieve the secret from Azure Key Vault
 				KeyVaultSecret secret = await client.GetSecretAsync(secretName);
-				LogPanel.Add(secret.Value);
+				LogPanel.Add($"{secret.Value}\r\n");
 				return secret.Value;
 			}
 			catch (Exception ex)
 			{
-				LogPanel.Add(ex.ToString());
+				LogPanel.Add(ex.ToString() + "\r\n");
 			}
 			return null;
 		}
@@ -231,10 +119,10 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 		/// <returns></returns>
 		public async Task<string> GetSecretFromKeyVaultAsyncUseProfile()
 		{
-			var profileResult = GetProfile();
+			var profileResult = Global.Security.GetProfile();
 			if (!profileResult.Success)
 			{
-				LogPanel.Add(string.Join("\r\n", profileResult.Errors));
+				LogPanel.Add(string.Join("\r\n", profileResult.Errors) + "\r\n");
 				return null;
 			}
 			var keyVaultName = KeyVaultNameTextBox.Text;
@@ -250,12 +138,12 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 
 				// Retrieve the secret from Azure Key Vault
 				KeyVaultSecret secret = await client.GetSecretAsync(secretName);
-				LogPanel.Add(secret.Value);
+				LogPanel.Add($"{secret.Value}\r\n");
 				return secret.Value;
 			}
 			catch (Exception ex)
 			{
-				LogPanel.Add(ex.ToString());
+				LogPanel.Add($"{ex}\r\n");
 			}
 			return null;
 		}
@@ -263,13 +151,57 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 
 		private void MainTabControl_SelectionChanged(object sender, SelectionChangedEventArgs e)
 		{
-
 		}
 
 		private async void TestButton_Click(object sender, RoutedEventArgs e)
 		{
+			LogPanel.Clear();
+			var profileResult = Global.Security.GetProfile();
+			if (profileResult.Success)
+				InspectToken(profileResult.Result.AccessToken);
 			await FetchAzureInformation();
 		}
+
+
+		public async Task LoadUserProfile()
+		{
+			// Load saved user profile
+			var profileResult = Global.Security.GetProfile();
+			if (!profileResult.Success)
+			{
+				LogPanel.Add(string.Join("\r\n", profileResult.Errors) + "\r\n");
+				return;
+			}
+			var profile = profileResult.Result;
+			UserName.Text = profile.Username;
+			UserAvatar.Source = await Global.Security.GetUserAvatar(profile.AccessToken);
+		}
+
+		public void InspectToken(string token)
+		{
+			var handler = new JwtSecurityTokenHandler();
+			try
+			{
+				var jwtToken = handler.ReadToken(token) as JwtSecurityToken;
+				if (jwtToken == null)
+				{
+					LogPanel.Add("Invalid JWT token.\r\n");
+					return;
+				}
+				// Display the expiry date
+				var expiryDate = jwtToken.ValidTo;
+				LogPanel.Add("Token Expiry Date: {expiryDate}\r\n");
+				// Optionally, you can print other claims as well
+				foreach (var claim in jwtToken.Claims)
+					Console.WriteLine($"Claim Type: {claim.Type}, Claim Value: {claim.Value}\r\n");
+			}
+			catch (Exception ex)
+			{
+				LogPanel.Add($"{ex}\r\n");
+			}
+		}
+
+
 	}
 
 }