+++ Engine/Security/AccessTokenCredential.cs
@@ -3,7 +3,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace JocysCom.VS.AiCompanion.Engine
+namespace JocysCom.VS.AiCompanion.Engine.Security
 {
 	public class AccessTokenCredential : TokenCredential
 	{